+++ phantom-dsl/src/test/scala/com/websudos/phantom/builder/primitives/PrimitivesTest.scala
@@ -53,7 +53,7 @@ class PrimitivesTest extends FlatSpec with Matchers {
     Primitive[ByteBuffer].asCql(slice) shouldEqual "0x030405"
   }
 
-  it should "autogenerate list primitives for known primitive types" in {
+  it should "autogenerate list primitives for List types" in {
     val test = Primitive[List[String]]
   }
